Shuffles around files. Move functions to methods.
Tests are passing and server appearing to give accurate responses.

Can tidy this up further later.

But this feels good enough for now. So will probably stay this way for a
while.
