+++ Cargo.toml
@@ -12,6 +12,7 @@ serde = { version = "1.0.210", features = ["derive"] }
 rayon = "1.10"
 actix-web = "4.0"
 env_logger = "0.11"
+awc = "3.5.1"
 
 [dev-dependencies]
 assert_approx_eq = "1.0"
