+++ src/core/mod.rs
@@ -1,0 +1,3 @@
+pub mod points;
+pub mod schedules;
+pub mod segment;
