+++ src/core/schedules/mod.rs
@@ -1,0 +1,2 @@
+pub mod amount_schedule;
+pub mod marginal_schedule;
