+++ src/lib.rs
@@ -1,0 +1,5 @@
+pub mod controller;
+pub mod core;
+pub mod errors;
+pub mod exchange_rates;
+pub mod utils;
