--- src/models/mod.rs
@@ -1,5 +1,0 @@
-pub mod request;
-pub mod response;
-pub mod segment;
-pub mod tax_schedule;
-pub mod taxes_config;
