--- src/models/tax_schedule.rs
@@ -1,534 +1,0 @@
-use crate::generate_range;
-use crate::Deserialize;
-use crate::IncomeTaxKnot;
-use crate::IncomeTaxPoint;
-use crate::LinearPiecewiseSegment;
-use crate::MarginalRateKnot;
-use crate::TaxError;
-use rayon::prelude::*;
-
-/// TODO: Need to split out the representations
-/// An income tax table is represented here
-#[derive(Deserialize, Debug, Clone)]
-pub struct IncomeTaxSchedule {
-    /// A sorted vector of points where the marginal tax rates change.
-    pub schedule: Vec<MarginalRateKnot>,
-}
-impl IncomeTaxSchedule {
-    // Lots to clean up here. Might be a need to have two different schedule classes to ensure no
-    // confusion
-
-    /// Given income tax knots, do binary search to find the segment and then interpolate
-    pub fn compute_income_tax(income: f32, income_tax_knots: &Vec<IncomeTaxKnot>) -> Option<f32> {
-        // TODO: This doesn't catch all the edge cases but should be good enough for now.
-        if income < 0.0 {
-            return None;
-        }
-        let mut l = 0;
-        let mut r = income_tax_knots.len();
-        while l < r {
-            // Cut down search space
-            let mid = l + (r - l) / 2;
-            if income_tax_knots[mid].income_limit == income {
-                return Some(income_tax_knots[mid].income_tax_amount);
-            }
-            if income_tax_knots[mid].income_limit < income {
-                if income_tax_knots[l].income_limit <= income
-                    && income <= income_tax_knots[l + 1].income_limit
-                {
-                    return LinearPiecewiseSegment {
-                        left_point: IncomeTaxKnot {
-                            income_limit: income_tax_knots[l].income_limit,
-                            income_tax_amount: income_tax_knots[l].income_tax_amount,
-                        },
-                        right_point: IncomeTaxKnot {
-                            income_limit: income_tax_knots[l + 1].income_limit,
-                            income_tax_amount: income_tax_knots[l + 1].income_tax_amount,
-                        },
-                    }
-                    .linear_interpolation(income);
-                }
-                l = mid;
-            } else {
-                if income_tax_knots[r - 1].income_limit <= income
-                    && income <= income_tax_knots[r].income_limit
-                {
-                    return LinearPiecewiseSegment {
-                        left_point: IncomeTaxKnot {
-                            income_limit: income_tax_knots[r - 1].income_limit,
-                            income_tax_amount: income_tax_knots[r - 1].income_tax_amount,
-                        },
-                        right_point: IncomeTaxKnot {
-                            income_limit: income_tax_knots[r].income_limit,
-                            income_tax_amount: income_tax_knots[r].income_tax_amount,
-                        },
-                    }
-                    .linear_interpolation(income);
-                }
-                r = mid;
-            }
-        }
-        None
-    }
-    /// Dot((r_i - r_{i-1}), max(0, x - b_{i-1}) where (b_0, r_0) = (0,0)
-    fn get_tax_amount_from_marginal_rates_knots(&self, income: f32) -> Result<f32, TaxError> {
-        if income < 0.0 {
-            return Err(TaxError::NegativeIncome);
-        }
-        let marginal_tax_rates_knots = &self.schedule;
-        let mut tax_amount = 0.0;
-        for (i, marginal_tax_knot) in marginal_tax_rates_knots.iter().enumerate() {
-            let prev_limit = if i > 0 {
-                marginal_tax_rates_knots[i - 1].income_limit
-            } else {
-                Some(0.0)
-            };
-            let prev_rate = if i > 0 {
-                marginal_tax_rates_knots[i - 1].marginal_rate
-            } else {
-                0.0
-            };
-            tax_amount += (marginal_tax_knot.marginal_rate - prev_rate)
-                * (income - prev_limit.expect("Error")).max(0.0);
-        }
-        Ok(tax_amount)
-    }
-
-    /// Better representation for efficient computation of taxes
-    pub fn to_income_amount_schedule(&self, max_income_to_consider: f32) -> Vec<IncomeTaxKnot> {
-        // Needs (0,0)
-        let mut income_tax_knots = vec![IncomeTaxKnot {
-            income_limit: 0.0,
-            income_tax_amount: 0.0,
-        }];
-        for (i, marginal_rate_knot) in self.schedule.iter().enumerate() {
-            if i == self.schedule.len() - 1
-                || marginal_rate_knot.income_limit.unwrap() > max_income_to_consider
-            {
-                break; // skip the last knot and replace with the max income to consider. Or truncate early.
-            }
-            income_tax_knots.push(IncomeTaxKnot {
-                income_limit: marginal_rate_knot.income_limit.expect("Error"),
-                income_tax_amount: self
-                    .get_tax_amount_from_marginal_rates_knots(
-                        marginal_rate_knot.income_limit.expect("Error"),
-                    )
-                    .expect("Error"),
-            });
-        }
-        income_tax_knots.push(IncomeTaxKnot {
-            income_limit: max_income_to_consider,
-            income_tax_amount: self
-                .get_tax_amount_from_marginal_rates_knots(max_income_to_consider)
-                .expect("Error"),
-        });
-        income_tax_knots
-    }
-}
-
-fn group_incomes_by_segment(
-    incomes: &[f32],
-    knot_points: &[IncomeTaxKnot],
-) -> Vec<(LinearPiecewiseSegment, Vec<f32>)> {
-    // Small bit of inefficiency with segments representation as it doubles up e.g. r1 == l2
-    // But this should not matter since number of knot points is usually low.
-    // Using LinearPiecewiseSegment makes things more readable too.
-    // This can be improved by avoiding cloning.
-    let mut point_index = 0;
-    let mut incomes_in_segment = Vec::new();
-    let mut overall_result = Vec::new();
-    let mut income_index = 0;
-    while income_index < incomes.len() {
-        let income = incomes[income_index];
-        if point_index + 1 >= knot_points.len() {
-            break;
-        }
-        // Assumes sorted incomes and sorted knot points.
-        if income <= knot_points[point_index + 1].income_limit {
-            incomes_in_segment.push(income);
-            income_index += 1;
-        } else {
-            overall_result.push((
-                LinearPiecewiseSegment {
-                    left_point: knot_points[point_index].clone(),
-                    right_point: knot_points[point_index + 1].clone(),
-                },
-                incomes_in_segment.clone(),
-            ));
-            incomes_in_segment.clear();
-            point_index += 1;
-        }
-    }
-
-    // Handle the final segment (wont be flushed otherwise...)
-    if !incomes_in_segment.is_empty() && point_index < knot_points.len() - 1 {
-        overall_result.push((
-            LinearPiecewiseSegment {
-                left_point: knot_points[point_index].clone(),
-                right_point: knot_points[point_index + 1].clone(),
-            },
-            incomes_in_segment,
-        ));
-    }
-
-    overall_result
-}
-
-/// Compute income tax amounts for a range of incomes
-fn compute_income_taxes_efficiently(
-    income_start: f32,
-    income_stop: f32,
-    income_step: f32,
-    income_tax_knot_points: &[IncomeTaxKnot],
-) -> Result<Vec<f32>, TaxError> {
-    // Not tested yet
-    let incomes_to_compute = generate_range(income_start, income_stop, income_step);
-    interpolate_segments_parallel(&incomes_to_compute, &income_tax_knot_points)
-}
-
-// Efficiently interpolate segments for a range of incomes
-pub fn interpolate_segments_parallel(
-    incomes: &[f32],
-    income_tax_knot_points: &[IncomeTaxKnot],
-) -> Result<Vec<f32>, TaxError> {
-    if incomes.last().unwrap() > &income_tax_knot_points.last().unwrap().income_limit {
-        return Err(TaxError::IncomeOutOfBounds);
-    }
-    // Choose not to parallelise the segments because the number of segments are usually low.
-    let grouped_income_values = group_incomes_by_segment(&incomes, &income_tax_knot_points);
-    Ok(grouped_income_values
-        .par_iter()
-        .flat_map(|(segment, income_group)| {
-            income_group
-                .par_iter()
-                .map(|&income| segment.linear_interpolation(income).unwrap())
-        })
-        .collect())
-}
-
-pub fn compute_breakeven_points(
-    knots1: &[IncomeTaxKnot],
-    knots2: &[IncomeTaxKnot],
-) -> Vec<IncomeTaxPoint> {
-    // Do not forget that for knots derived from tax schedules we need to define an upper bound (cannot be inf, inf...)
-
-    let mut i = 0; // curve 1
-    let mut j = 0; // curve 2
-    let mut breakeven_points = Vec::new();
-
-    while i + 1 < knots1.len() && j + 1 < knots2.len() {
-        let l1 = knots1[i].income_limit;
-        let r1 = knots1[i + 1].income_limit;
-        let l2 = knots2[j].income_limit;
-        let r2 = knots2[j + 1].income_limit;
-
-        let segments_have_overlap = r1 >= l2 && r2 >= l1;
-
-        if segments_have_overlap {
-            let candidate_segment = LinearPiecewiseSegment {
-                left_point: knots2[j].clone(),
-                right_point: knots2[j + 1].clone(),
-            };
-            let intersection = LinearPiecewiseSegment {
-                left_point: knots1[i].clone(),
-                right_point: knots1[i + 1].clone(),
-            }
-            .compute_intersection(&candidate_segment);
-            match intersection {
-                None => {} // do nothing
-                Some(breakeven_point) => {
-                    breakeven_points.push(breakeven_point); // add to result set
-                }
-            };
-        }
-
-        if r1 < r2 {
-            i += 1; // advance 1
-        } else {
-            j += 1; // advance 2
-        }
-    }
-    breakeven_points
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::models::tax_schedule::*;
-    use crate::utils::utils::income_points_are_approx_eq;
-    use crate::{IncomeTaxKnot, MarginalRateKnot, TaxError};
-    #[test]
-    fn test_marginal_rates_schedule_to_income_tax_amount_schedule() {
-        let schedule = IncomeTaxSchedule {
-            schedule: vec![
-                MarginalRateKnot {
-                    marginal_rate: 0.1,
-                    income_limit: Some(10000.0),
-                },
-                MarginalRateKnot {
-                    marginal_rate: 0.2,
-                    income_limit: Some(20000.0),
-                },
-                MarginalRateKnot {
-                    marginal_rate: 0.3,
-                    income_limit: Some(f32::INFINITY),
-                },
-            ],
-        };
-
-        let max_income_to_consider = 100000.0;
-        let expected_result = vec![
-            IncomeTaxKnot {
-                income_limit: 0.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 10000.0,
-                income_tax_amount: 1000.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 20000.0,
-                income_tax_amount: 3000.0,
-            },
-            IncomeTaxKnot {
-                income_limit: max_income_to_consider,
-                income_tax_amount: 27000.0,
-            },
-        ];
-
-        let actual_result = schedule.to_income_amount_schedule(max_income_to_consider);
-        assert_eq!(expected_result, actual_result);
-    }
-
-    #[test]
-    fn test_get_tax_amounts_from_tax_amounts_schedule() {
-        // Using example from wikipedia: https://en.wikipedia.org/wiki/Progressive_tax
-        let income_tax_knots = vec![
-            IncomeTaxKnot {
-                // required lower bound must be injected in the conversion.
-                income_tax_amount: 0.0,
-                income_limit: 0.0,
-            },
-            IncomeTaxKnot {
-                income_tax_amount: 1000.0,
-                income_limit: 10000.0,
-            },
-            IncomeTaxKnot {
-                income_tax_amount: 3000.0,
-                income_limit: 20000.0,
-            },
-            // An explicit upper bound must be defined in this format.
-            // This is known as the "max income to consider"
-            IncomeTaxKnot {
-                income_tax_amount: 27000.0,
-                income_limit: 100000.0,
-            },
-        ];
-
-        let result = IncomeTaxSchedule::compute_income_tax(25000.0, &income_tax_knots);
-        assert_eq!(result, Some(4500.0));
-
-        let result = IncomeTaxSchedule::compute_income_tax(5000.0, &income_tax_knots);
-        assert_eq!(result, Some(500.0));
-
-        let invalid_result = IncomeTaxSchedule::compute_income_tax(-25000.0, &income_tax_knots);
-        assert!(invalid_result.is_none());
-
-        let zero_result = IncomeTaxSchedule::compute_income_tax(0.0, &income_tax_knots);
-        assert_eq!(zero_result, Some(0.0));
-    }
-
-    #[test]
-    fn test_get_tax_amounts_from_marginal_tax_rates_schedule() {
-        // Using example from wikipedia: https://en.wikipedia.org/wiki/Progressive_tax
-        let schedule = IncomeTaxSchedule {
-            schedule: vec![
-                MarginalRateKnot {
-                    marginal_rate: 0.1,
-                    income_limit: Some(10000.0),
-                },
-                MarginalRateKnot {
-                    marginal_rate: 0.2,
-                    income_limit: Some(20000.0),
-                },
-                MarginalRateKnot {
-                    marginal_rate: 0.3,
-                    income_limit: Some(f32::INFINITY),
-                },
-            ],
-        };
-
-        let result = schedule.get_tax_amount_from_marginal_rates_knots(25000.0);
-        assert_eq!(result.unwrap(), 4500.0);
-
-        let invalid_result = schedule.get_tax_amount_from_marginal_rates_knots(-25000.0);
-        assert_eq!(invalid_result.unwrap_err(), TaxError::NegativeIncome);
-
-        let zero_result = schedule.get_tax_amount_from_marginal_rates_knots(0.0);
-        assert_eq!(zero_result.unwrap(), 0.0);
-    }
-
-    #[test]
-    fn test_parallel_interpolate() {
-        let incomes = vec![500.0, 1500.0, 1700.0, 2500.0];
-        // not upper bounded.
-        let knot_points = vec![
-            IncomeTaxKnot {
-                // always need a (0,0) for it to work.
-                income_limit: 0.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 1000.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 2000.0,
-                income_tax_amount: 1.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 3000.0,
-                income_tax_amount: 3.0,
-            },
-        ];
-        let actual_result = interpolate_segments_parallel(&incomes, &knot_points);
-        let expected_result = vec![
-            0.0, 0.5, 0.7, // y = (1/1000)*(x - 2000) + 1
-            2.0,
-        ];
-        assert_eq!(actual_result, Ok(expected_result));
-
-        // it should throw error because income is out of bounds w.r.t. the tax schedule
-        let invalid_incomes = vec![500.0, 1500.0, 1700.0, 2500.0, 3500.0];
-        let invalid_result = interpolate_segments_parallel(&invalid_incomes, &knot_points);
-        assert_eq!(invalid_result.unwrap_err(), TaxError::IncomeOutOfBounds);
-    }
-
-    #[test]
-    fn test_group_incomes_by_segment() {
-        let incomes = vec![500.0, 1500.0, 1700.0, 2500.0, 3500.0];
-        // is not upper bounded.
-        let knot_points = vec![
-            IncomeTaxKnot {
-                // always need a (0,0) for it to work.
-                income_limit: 0.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 1000.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 2000.0,
-                income_tax_amount: 1.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 3000.0,
-                income_tax_amount: 3.0,
-            },
-        ];
-
-        let expected_result = vec![
-            (
-                LinearPiecewiseSegment {
-                    left_point: IncomeTaxKnot {
-                        income_limit: 0.0,
-                        income_tax_amount: 0.0,
-                    },
-                    right_point: IncomeTaxKnot {
-                        income_limit: 1000.0,
-                        income_tax_amount: 0.0,
-                    },
-                },
-                vec![500.0],
-            ),
-            (
-                LinearPiecewiseSegment {
-                    left_point: IncomeTaxKnot {
-                        income_limit: 1000.0,
-                        income_tax_amount: 0.0,
-                    },
-                    right_point: IncomeTaxKnot {
-                        income_limit: 2000.0,
-                        income_tax_amount: 1.0,
-                    },
-                },
-                vec![1500.0, 1700.0],
-            ),
-            (
-                LinearPiecewiseSegment {
-                    left_point: IncomeTaxKnot {
-                        income_limit: 2000.0,
-                        income_tax_amount: 1.0,
-                    },
-                    right_point: IncomeTaxKnot {
-                        income_limit: 3000.0,
-                        income_tax_amount: 3.0,
-                    },
-                },
-                vec![2500.0],
-            ),
-        ];
-
-        let actual_result = group_incomes_by_segment(&incomes, &knot_points);
-        assert_eq!(expected_result, actual_result);
-    }
-
-    #[test]
-    fn test_get_breakeven_points() {
-        // https://www.desmos.com/calculator
-        // Breakevens at (25/3, 20/3) and (50/3, 40/3)
-
-        let curve1 = vec![
-            IncomeTaxKnot {
-                income_limit: 0.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 5.0,
-                income_tax_amount: 5.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 15.0,
-                income_tax_amount: 10.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 20.0,
-                income_tax_amount: 20.0,
-            },
-        ];
-        let curve2 = vec![
-            IncomeTaxKnot {
-                income_limit: 5.0,
-                income_tax_amount: 0.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 10.0,
-                income_tax_amount: 10.0,
-            },
-            IncomeTaxKnot {
-                income_limit: 20.0,
-                income_tax_amount: 15.0,
-            },
-        ];
-
-        let breakevens = compute_breakeven_points(&curve1, &curve2);
-        let tolerance = 1e-5;
-        assert!(income_points_are_approx_eq(
-            breakevens[0].clone(),
-            IncomeTaxPoint {
-                income: 25.0 / 3.0,
-                income_tax_amount: 20.0 / 3.0
-            },
-            tolerance
-        ));
-
-        assert!(income_points_are_approx_eq(
-            breakevens[1].clone(),
-            IncomeTaxPoint {
-                income: 50.0 / 3.0,
-                income_tax_amount: 40.0 / 3.0
-            },
-            tolerance
-        ));
-    }
-}
